Creativity Challenge
In this challenge, a creativity task was created. This task consisted of the robot using ultrasonic sensor and when the robot is in moving motion, there are green LEDs mounted on it that light up and as it gets closer to objects, the green and red LEDs start blinking. Once it is 2 inches close to the wall, the robot stops and the red LED lights up. Then the robot goes backwards and makes a turn to right and restarts the process again.